Add support for the debug runtime in MSVC builds.
We can force the linker to choose the debug runtime, which is necessary
to override rusts libc, which by default wants to link to the release
runtime, even in debug mode.

With this patch CMake will override this behavior and use the debug
runtime if appropriate.
This is very important, as all parts of the program must agree on the runtime!
As Qt in Debug mode uses the debug runtime in their DLLs, we may
encounter undefined behavior otherwise.